moving the template utility to a shared space
