+++ src/core/markdown/adapters/renderable-to-page-data.ts
@@ -1,7 +1,7 @@
 import { ReferenceGuideReference, Renderable, RenderableBundle, RenderableEnum } from './types';
 import { DocPageData, DocumentationBundle } from '../../shared/types';
 import { pipe } from 'fp-ts/function';
-import { CompilationRequest, Template } from '../templates/template';
+import { CompilationRequest, Template } from '../../template';
 import { enumMarkdownTemplate } from '../templates/enum-template';
 import { interfaceMarkdownTemplate } from '../templates/interface-template';
 import { classMarkdownTemplate } from '../templates/class-template';
